Final Push
This push has an improved model for training the data. The loss was brought from 0.30 to 0.06 with this push. This push also includes an improvement where the confusion matrices are now saved as.png files in the "Confusion_Matrices" folder. 

Small improvements to the program could be made, but overall it is very functional.